Fix remote/branch splitting and support branches with "/" in their name
For some odd reason, support for branches with "/" in their name was
disabled in commit c43d4c160589 ("Check for backslashes in remote
branch name"), instead of fixing the real problems.

Running the status command that the Load function runs, manually, I
get:

Load function uses "git status -b --porcelain -u no" to retrieve the
remote name, and the remote branch name.  It incorrectly parses the
case of branches with "/", however.  Running the status command
manually, I get:

  $ git status -b --porcelain -u no
  ## palves/spr-test...github_palves/palves/spr-test [ahead 1]

And Load parses it as:

  GitHubRemote = "github_palves/palves"
  GitHubBranch = "spr-test"

While it should instead be:

  GitHubRemote = "github_palves"
  GitHubBranch = "palves/spr-test"

This is a classic greedy regex problem: the current regexp's second
capture (([a-zA-Z0-9_\-/\.]+)) grabs as much as it can while still
leaving something for the final /... group, so it steals palves and
leaves only spr-test for the last group.

The fix here is to split the remote reference at the first / (the
boundary between remote name and remote branch).  That can be done
either by making the regex stop at the first /, or by capturing the
whole remote-ref and splitting it with strings.SplitN.

I'm doing the latter (capture full remote-ref and SplitN), as I find
it simpler and to the point and thus more robust.

(cherry picked from commit 30e8ffbea18bebd1e1706fadc3047ba9fea08883)

v2 (Nicolai Hähnle <nhaehnle@gmail.com>):
- formatting