Add a "[N/M]" prefix to the stack listing in the PR description
This is inspired by commit 0b4f7a014e ("Add a "[<series name> N/M] "
prefix to the title of each PR in the stack") by Pedro Alves.

Adding it to the PR description makes it more intuitive for people to
understand the order in which the stack is displayed (which matches a
`git log` but not the order in which you might see a patch series in an
email client). *Not* adding it to the PR title seems less disruptive to
teams not used to stacked PRs (though perhaps it should be available
under a config flag).